clean up, route naming
