+++ app/(dashboard)/repo-settings/[repositoryId]/page.tsx
@@ -1,0 +1,5 @@
+import { redirect } from "next/navigation";
+
+export default function RepoSettingsDashboard({ params }) {
+  return redirect(`/repo-settings/${params.repositoryId}/players`);
+}
